94_binary_tree_inorder_traversal/main_test.go
package main

import (
	"testing"

	"github.com/stretchr/testify/require"
)

func TestInorderTraversal(t *testing.T) {
	tests := []struct {
		name     string
		root     *TreeNode
		expected []int
	}{
		{
			name: "partly",
			root: &TreeNode{
				Val:  1,
				Left: nil,
				Right: &TreeNode{
					Val: 2,
					Left: &TreeNode{
						Val:   3,
						Left:  nil,
						Right: nil,
					},
					Right: nil,
				},
			},
			expected: []int{1, 3, 2},
		},
		{
			name:     "empty",
			root:     nil,
			expected: []int{},
		},
		{
			name: "only root",
			root: &TreeNode{
				Val:   1,
				Left:  nil,
				Right: nil,
			},
			expected: []int{1},
		},
		{
			name: "full",
			root: &TreeNode{
				Val: 1,
				Left: &TreeNode{
					Val: 2,
					Left: &TreeNode{
						Val:   3,
						Left:  nil,
						Right: nil,
					},
					Right: &TreeNode{
						Val:   4,
						Left:  nil,
						Right: nil,
					},
				},
				Right: &TreeNode{
					Val: 5,
					Left: &TreeNode{
						Val:   6,
						Left:  nil,
						Right: nil,
					},
					Right: &TreeNode{
						Val:   7,
						Left:  nil,
						Right: nil,
					},
				},
			},
			expected: []int{3, 2, 4, 1, 6, 5, 7},
		},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			require.Equal(t, tt.expected, inorderTraversal(tt.root))
		})
	}
}
